requirements.txt
colorama==0.4.4
Pillow==8.0.1
qrcode==6.1
six==1.15.0
